Patch - Bug Fixes - Null pointer exception on deleting app solved.

+++ src/main/java/com/pezy/pezy_api/helper/FirebaseDatabaseHelper.java
@@ -73,7 +73,9 @@ public class FirebaseDatabaseHelper {
 			System.out.println(msg.getMessage());
 		}
 		msg.setStatus(false);
-		apps.delete();
+		if(null != apps) {
+			apps.delete();
+		}
 		return msg;
 	}
